@@ -3,12 +3,11 @@
 apply from: file("../Android_Lib/lib_sdk/config.gradle")
 
 buildscript {
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '1.5.20'
 //    ext.gradle_version = "4.0.1" // Entry name 'res/drawable/notification_bg_low.xml' collided
     ext.greendao_version = "3.3.0"
-    ext.butterknife_version = '10.1.0'
 
-    ext.gradle_version = "3.5.0"
+    ext.gradle_version = "4.0.1"
 
     repositories {
         google()
@@ -17,10 +16,9 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:${gradle_version}"
+        classpath "com.android.tools.build:gradle:$gradle_version"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${kotlin_version}"
         classpath "org.greenrobot:greendao-gradle-plugin:${greendao_version}"
-        classpath "com.jakewharton:butterknife-gradle-plugin:${butterknife_version}"   // 添加
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
