@@ -2,32 +2,22 @@ package holder;
 
 import android.content.Context;
 import android.view.View;
-import android.widget.TextView;
+
+import androidx.databinding.DataBindingUtil;
 
 import com.as_160213_bd_map.R;
+import com.as_160213_bd_map.databinding.OfflineLocalmapItemBinding;
 import com.baidu.mapapi.map.offline.MKOLUpdateElement;
 import com.baidu.mapapi.map.offline.MKOfflineMap;
 
 import adp.LocalMapAdp;
 import base.BaseHolder;
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
 import lib.toast.XToast;
 
 
 public class LocalMapHolder extends BaseHolder {
 
-    private Context ctx;
-
-    @BindView(R.id.offline_localMap_title)
-    TextView offline_localMap_title;
-
-    @BindView(R.id.offline_localMap_update)
-    TextView offline_localMap_update;
-
-    @BindView(R.id.offline_localMap_ratio)
-    TextView offline_localMap_ratio;
+    private OfflineLocalmapItemBinding offlineLocalmapItemBinding;
 
     //
     private MKOLUpdateElement e;
@@ -39,22 +29,14 @@ public class LocalMapHolder extends BaseHolder {
     public LocalMapHolder(Context ctx) {
         super(ctx);
 
-        this.ctx = ctx;
     }
 
     @Override
     public View initConvertView(Context ctx) {
         View v = View.inflate(ctx, R.layout.offline_localmap_item, null);
-
-        ButterKnife.bind(this, v);
-        return v;
-    }
-
-
-    @OnClick({R.id.offline_localMap_remove})
-    void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.offline_localMap_remove:
+        offlineLocalmapItemBinding = DataBindingUtil.bind(v);
+        if (null != offlineLocalmapItemBinding) {
+            offlineLocalmapItemBinding.offlineLocalMapRemove.setOnClickListener(v1 -> {
                 boolean result = mkOfflineMap.remove(e.cityID);
                 if (result) {
                     if (callBack != null) {
@@ -62,22 +44,24 @@ public class LocalMapHolder extends BaseHolder {
                     }
                     XToast.INSTANCE.show("Delete Success");
                 }
-                break;
+            });
         }
+        return v;
     }
+
 
     public void setData(MKOLUpdateElement e, MKOfflineMap mkOfflineMap, LocalMapAdp.CallBack callBack) {
         this.e = e;
         this.mkOfflineMap = mkOfflineMap;
         this.callBack = callBack;
 
-        offline_localMap_title.setText(e.cityName);
+        offlineLocalmapItemBinding.offlineLocalMapTitle.setText(e.cityName);
         String radio = e.ratio + "%";
-        offline_localMap_ratio.setText(radio);
+        offlineLocalmapItemBinding.offlineLocalMapRatio.setText(radio);
         if (e.update) {
-            offline_localMap_update.setText("可更新");
+            offlineLocalmapItemBinding.offlineLocalMapUpdate.setText("可更新");
         } else {
-            offline_localMap_update.setText("最新");
+            offlineLocalmapItemBinding.offlineLocalMapUpdate.setText("最新");
         }
     }
 
